add product sku add 1,2,3,4 in end and selling price float

+++ client/src/views/AdminProduct.jsx
@@ -95,7 +95,7 @@ const AdminProduct = () => {
     switch (ref) {
       case 'price':
         newArr[index].price = e.target.value;
-        newArr[index].selliingPrice = parseInt(e.target.value) + ((parseInt(e.target.value) * parseInt(autoMargin))/100);
+        newArr[index].selliingPrice = parseFloat(e.target.value) + ((parseFloat(e.target.value) * parseFloat(autoMargin))/100);
         break;
       case 'quantity':
         newArr[index].quantity = e.target.value;
